feat: create mock Vertex core setup (#3)
**Motivation:**

This includes the Router (Governor) that supports a single executor, a
basic executor, a mock voting NFT, and a mock protocol.

**Modifications:**

Created the router, executor, and mocks. Also configured the test setup
to test the propose and execute process.

**Result:**

This will serve as a strong foundation for adapting OZ Governor into the
full Vertex vision.