refactor: small fixes to existing roles implementation (#186)
**Motivation:**

Another small PR, pre-bigger refactor, to fix some issues and clean
things up

**Modifications:**

- All methods that change role holders assert that the supply of admin
roles is nonzero at the end of the call
- Dedupe custom errors and simplify usage by including the role number
in the error
- Fix bug where all role holders quantities were being incremented by
`quantity` and not by 1

**Result:**

This is the final cleanup PR to make sure the larger refactor PR that's
coming next is easy to review. The next PR will:
- Require initializing a role with a description string before using it
on-chain. This ensures the contract knows how many roles exist (so we
can enumerate them during revocation of policies), and it allows us to
tie role IDs to description strings
- Add a way to batch calls to multiple policy methods into a single
action