test: finish factory and core invariant tests (#212)
**Motivation:**

As part of #52, this finishes invariant tests for the factory. There is
only 1 invariant I can think of, so that's all we're testing. If you
think of more let me know.

This PR also begins scaffolding out the VertexCore invariant tests

**Modifications:**

- Adds new factory methods to the factory handler contract
- Cleans up imports in the factory and the invariant test file
- Re-ordered the methods in the factory so the public view function
comes after all state changing methods
- Clarified a few comments in `VertexStrategy`
- Fix the name of the `VertexCore` invariant test file
- Setup `VertexCore` invariant tests.

**Result:**

Invariant tests for the factory are complete, invariant tests for the
core are scaffolded