feat: Deploy script (#215)
**Motivation:**

We want an automated way to deploy the Vertex infrastructure on
different networks, ensuring that important contracts on each network
have the same address. We also want to make sure that this is done
safely, ideally without the possibility that anyone could frontrun a
transaction, e.g. to intialize an important contract.

**Modifications:**

This PR:
* adds a `DeployVertex` script to deploy logic contracts, the lens, and
the factory to a new chain
* adds a README documenting safe script usage and best practices for use
at deploy time
* uses `DeployVertex` as a part of the test suite setup, so that all
tests run against the state prepared by the deploy script

**Result:**

We have a script to safely deploy Vertex on multiple chains