refactor: ignore creator in strategy quorums (#261)
**Motivation:**

EDIT: Leaving this text for posterity sake but decided to only disallow
creators from participating in absolute strategies and allow creators to
participate in relative strategies.

Closes #256. This adjusts how we handle approval and disapproval
accounting logic to ignore the action creators. When an action creator
has an approvalRole or disapprovalRole of the action they created, then
they should not be able to participate in the strategy process. We can
remove their ability to approve or disapprove but we should also update
how quorum is reached. My approach here is as follows:

**Relative Strategy**

- Check if the action creator has the approvalRole, if they do reduce
`approvalPolicySupply` by 1
- Check if the action creator has the disapprovalRole, if they do reduce
`disapprovalPolicySupply` by 1

Since relative strategies are percentage based the approval and
disapproval thresholds will update correctly by making this change to
the supplies.

**Absolute Strategy**

- Disallow action creation if `minApprovals` is greater than
`approvalPolicySupply - actionCreatorApprovalRoleQty`
- Disallow action creation if `minDisapprovals` is greater than
`disapprovalPolicySupply - actionCreatorDisapprovalRoleQty` (unless
minDisapprovals is `type(uint).max`)

This ensures that an action can't be created where the approval or
disapproval threshold are never met. This also means that
minDisapprovals must be `type(uint).max` to indicate there shouldn't be
a disapproval process.

**Modifications:**

Edit RelativeStrategy and AbsoluteStrategy to not include the creator in
the threshold calculations.

**Result:**

Relative Strategy;

Creator has role:
- Total policy supply is subtracted by 1 and captured at action creation
time. There are validation checks to ensure this doesn’t underflow.
Since the denominator is calculated using the total supply not quantity
and the threshold is percentage there isn’t any concern about the
threshold not being able to be met

Creator does not have role:
- Total policy supply (not quantity) is captured at action creation
time. The approval/disapproval threshold is this supply multiplied by
the min approval/disapproval percentage. If the quantity that casts an
approval/disapproval reaches this threshold then the action is
approved/disapproved.

Absolute Strategy:

Creator has role:
- No total supply calculation is made for absolute strategies. There is
a validation check to ensure that the threshold can still be met without
the action creator’s participation. Disabling the disapproval process is
done by setting the minDisapproval to `type(unit).max`.

Creator does not have role:
- There is a validation check to ensure that there is enough possible
approvals/disapprovals to meet thresholds. Disabling the disapproval
process is done by setting the minDisapproval to `type(unit).max`.