feat: check slot0 in account delegatecalls (#267)
**Motivation:**

Whenever we're executing arbitrary code in the context of an account
(i.e. a delegatecall), we want to ensure that none of the storage in
this contract changes, as this could let someone who sneaks in a
malicious (or buggy) target to take ownership of this contract. Slot 0
contains all relevant storage variables for security, so we check the
value before and after execution to make sure it's unchanged. The
contract name starts in slot 1, but it's not as important if that's
changed (and it can be changed back), so to save gas we don't check the
name. The storage layout of this contract is below, obtained with `forge
inspect LlamaAccount storage --pretty`:

```
| Variable Name | Type    | Slot | Offset | Bytes |
|---------------|---------|------|--------|-------|
| _initialized  | uint8   | 0    | 0      | 1     |
| _initializing | bool    | 0    | 1      | 1     |
| llamaCore     | address | 0    | 2      | 20    |
| name          | string  | 1    | 0      | 32    |
```
**Modifications:**

Add checks on the slot 0 value before and after the delegatecall

**Result:**

Malicious/buggy scripts cannot take ownership of the account