fix: dont allow new roles after action creation in same timestamp (#307)
**Motivation:**

https://github.com/spearbit-audits/review-llama/issues/23

**Modifications:**

- Adds a `getLastActionTimestamp` getter to simplify fetching the
timestamp of the most recently created action. This is specifically
needed by the policy, so I was on the fence about adding this bespoke
method for it (vs. having policy fetch last action count, then fetch the
action struct and extract the timestamp), but this will be cheaper and
there's not really a downside
- Added `_assertNoActionCreationsAtCurrentTimestamp` method that's
called from `_setRoleHolder`. The spearbit issue says we only need it
there, and after thinking about it I agree, but would be great to have
another sanity check there
- Sorted errors in `LlamaPolicy` alphabetically
- Adds a test for this new assertion

**Result:**

See above