feat: set color and logo on deploy for non-root instances (#327)
**Motivation:**

When a Llama instance is deployed its logo and color cannot be set. This
leads to a poor UX because instance's must permission the color and logo
setter to update this or we must do it for them. This will also lead NFT
marketplaces to cache incorrect versions of their policy.

**Modifications:**

Added a color and logo setter in the factory that is called when the
deploy function is called. I avoided adding this to the constructor
because I figure it would be nice for this to be an initial action for
the root llama instance but there are two things we could do here:

1. The constructor can accept a color and logo parameter to set for the
root llama instance.
2. We just hardcode our logo and color in the
`LlamaPolicyMetadataParamRegistry` constructor for the root instance.

**Result:**

Users will be able to atomically set their color and logo for their
policy NFT on deployment. This means a better UX for freshly deployed
Llama instances while minimizing protocol changes.