refactor: rename strategies (#339)
**Motivation:**

[This issue](https://github.com/spearbit-audits/review-llama/issues/38)
will be closed with two PRs: this one and a followup PR that addresses
https://github.com/llamaxyz/llama/issues/341.

**Modifications:**

We decided to address this finding by renaming `AbsoluteStrategy` to
`PeerReview`. In a follow-up PR I'll add a strategy that works the same
as `PeerReview` but will allow the action creator to cast. This strategy
will be called `AbsoluteQuorum`. `RelativeStrategy` was also renamed to
`RelativeQuorum`.

**Result:**

We can create a followup PR for the new strategy with a minimal diff.