refactor: remove `createAction` functions that don't have description param (#340)
**Motivation:**

Closes https://github.com/llamaxyz/llama/issues/336. To keep this
standard with other popular governance frameworks and remove unneeded
code, we're removing the `createAction` and `createActionBySig` that
doesn't have a description argument. If users don't want to provide a
description, they can pass an empty string.

**Modifications:**

Remove these two functions and the internal `_createActionBySig`
function. The rest of the PR is updating tests to pass an empty string
instead of not including the description argument.

**Result:**

`LlamaCore` will have less code and be more consistent with other
governance frameworks.