fix: set `ALL_HOLDERS_ROLE` expiration to 0 when policy is burnt (#407)
**Motivation:**

https://github.com/llamaxyz/llama/blob/main/src/LlamaPolicy.sol#L534

This should be `roleBalanceCkpts[tokenId][ALL_HOLDERS_ROLE].push(0,0);`
.

Currently when burning a token, the `ALL_HOLDERS_ROLE` will still have
an infinite expiry which is not consistent with the other roles which
require quantity and expiry to be both 0 when revoking roles.

Also making the above change, means that the `push()` function without
expiration in the `Checkpoints` lib is only used in the `_mint()`
function in LlamaPolicy. We can remove that function and use push with
expiration in the mint function to make it more explicit.

**Modifications:**

* Using `roleBalanceCkpts[tokenId][ALL_HOLDERS_ROLE].push(0, 0);` in
`_burn()` in `LlamaPolicy`
* Using `roleBalanceCkpts[tokenId][ALL_HOLDERS_ROLE].push(1,
type(uint64).max);` in `_mint()` in `LlamaPolicy`
* Appropriate test case (Also tested that it fails without the above
change)
* Removing `push()` function without expiration in `Checkpoints` lib

**Result:**

More consistent burning of a policy with all roles having their
quantities and expiration set to 0. Closes #406